refactor: refactor transfer_leader
