+++ server/cluster/node.go
@@ -9,6 +9,10 @@ type Node struct {
 	shardIDs []uint32
 }
 
+func (n Node) GetShardIDs() []uint32 {
+	return n.shardIDs
+}
+
 func (n Node) GetMeta() *clusterpb.Node {
 	return n.meta
 }
