+++ server/coordinator/procedure/dispatch/dispatch.go
@@ -1,0 +1,4 @@
+package dispatch
+
+type Dispatch interface {
+}
